level1c: Use limb_data recarray
Instead of declaring all limb data variables by themselves we combine
them into on numpy.recarray. This enables us to access the fields
dictionary-like or as attributes. It simplifies binary reading and
writing substantially but requires to fill the struct when reading from
one of the other file types.

Provides a convenience functions for to prepare the dtype and data
accordingly to the data size.

Provides a convenience function to read from HDF5 files such that it
provides the same interface as for .dat and. l_mpl_binary files.