l1c hdf5: Private members for radiances and errors
We are going to remove rad_list and err_list from the public
scia_limb_scan class members. This is the only place where they were
still used and made the HDF5 looping over the clusters rather simple.
To avoid too much rewriting, we replace them by private variables to be
filled and which are only used in this HDF5 reading part.