Regression models and MCMC sampling
Adds a new submodule for the regression models and MCMC sampling of
the posterior probability. The models use the `celerite`[1] modeling
protocol and treats the linear and non-linear case equally.
The models can also be used as mean models for Gaussian process
regression with `george`[2] or `celerite`.
MCMC sampling is done with `emcee`[3].

Includes a data loading interface for SCIAMACHY daily zonal mean
data (NO) and for solar and geomagnetic index files.

[1](https://celerite.readthedocs.io)
[2](https://george.readthedocs.io)
[3](https://emcee.readthedocs.io)