Proxy model: Allow centering the proxy values
In some cases it may be convenient to center the proxy values by
subtracting the mean of all values.

Before we subtracted the mean outside the class; including this logic in
the class itself helps to better keep track of this mean value. This
number is later needed to reconstruct the model predictions in the case
some outsider wants to use our parameters. In other words, he or she has
to subtract that number from his or her proxy values to be useful with
our fitted parameters.