lvl2: Include nrlmsise00 wrapper module
Wraps the NRLMSISE-00 C functions in a Python accessible way.

The input struct translates directly to a python dict, the flags'
switches are taken as a list of ints.

So far only gtd7() and gtd7d() are supported and the ouput is directly
tranferred to a python tuple of lists of doubles.