l2: SCIAMACHY retrieval results interfaces
Our home-grown classes to handle the SCIAMACHY level 2 results.
Started out as an interface for the plain text files, and evolved
to handle post-processed data with additional attributes/data.
Handles converting to/reading from netcdf files as well.

Nowadays we could use pandas and xarray, but we keep them to avoid
losing work. May vanish in the future.