l2 msis: Python 2 and old C-compiler compatibility
Apparently declaring the variables within for-loops is c99 or newer.
We could just add it to extra_compile_args in setup.py, but we fix it
here, increasing the chances that it then also works with very old
C compilers.

Python 2 doesn't know PyErr_WarnFormat() so we use PyErr_WarnEx() and
skip the flexible format. Still unsure if it would not be better to
raise an exception and error out on those cases.

We forgot to change the non-Python-3 (Python 2) module initialization
to our module name.