l2: Fix nrlmsise00 interface
We tried to propagate the python dictionary values for the inputs to the
nrlmsise00 input struct in a pure fashion, i.e. returning a new struct.
Since that struct includes a pointer to an array with ap values, the
location where it points to has to be valid. Pointing to a local
variable is not a good idea. The proper way would be to malloc() in
dict_get_double_default(), but then the caller(s) gtd7() and gtd7d()
would have to keep track of if and free() it.

Initializing the ap array on the caller side and then passing the
(reference to) the struct seems to be the simpler approach for now.