l2 nrlmsise00: Fix python 2/3 compatibility
Python 3 ints are long by default, but python 2 knows both long and int.
When checking passed values, PyInt_Check() is the way for py2, which
apparently works differently from its PyLong_Check().
But py3 only knows PyLong_Check() and to restore compatibility, we
define PyLong_Check() as PyInt_Check() for py3 and use that for
checking the values.