regress: Change kernel bounds setup
Instead of providing and setting the bounds separately, we directly set
them when we initialize the gaussian process kernels. Thus the bounds
now only appear in one place of the code, reducing the risk of confusing
people.

Propagates the kernel bounds to the MCMC handler since they are needed
for the latin hypercube seeding (LHS) of the walkers. LHS is used when
we use MCMC sampling also for the fit. Previously we could use that
initialization only for the celerite kernels. Now also the george
kernels are supported.