Celerite/george compatibility fixes
George's predict() takes the times as a mandatory second parameter where
it is an optional keyword argument for celerite's predict. When it is
not set, celerite uses the internally saved times for the data points.
We used predict() in a few places, mostly for plotting, which relied on
celerite's behaviour and broke when using george. We fix the times by
always explicitly setting the times via the keyword option which works
for both.

Another compatibility fix concerns the use of log_determinant when
calculating the diagnostics after the MCMC run. log_determinant is
a method in celerite.GP and a float in george.GP instances. We simply
try the method (celerite) variant first which raises a TypeError in the
case of george which we catch and use log_determinant as a float.
Related to the log_determinant we update the gaussian process internals
using compute() before calculating the diagnostics just to be sure
that everything is up to date according to the set parameters.