regress: Use all samples' log probabilities
Emcee saves the log probabilities of the chain which we can use instead
of the last state of the walkers. Using all probabilities should give a
better estimate of the maximum posterior solution and its value.

We can now correctly elevate the probabilities to the caller, for
example to plot the maximum posterior solution instead of the median
solution.