regress: Plot the maximum posterior solution
Enables the maximum posterior solution to be plotted similarly to the
median solution. Takes the returned probabilities from the MCMC sampler
which was fixed in commit bf232757c705c3ae5641e6c7ca663b55c6ae44dc.

Adds command line options for plotting the maxlnp solution and
residuals, defaults to plotting maxlnp now instead of the median.
Fixes plotting the solution at least in cases with the Matérn-3/2
kernel when the length scale distribution (log_rho) is more flat than
Gaussian and the median is large compared to the maximum.