regress: Change lag_times command line handling
We now prepare an all-zero lag dictionary and update the entries as they
are given on the command line. That way we can use lags and fit the
lifetime together, instead of forcing the lag to zero for the lifetime
fit.

Removes catching a possible ValueError so that the user knows that
something is wrong instead of doing something potentially unintended.

Updates the command line option defaults to not raise a ValueError when
no lags are specified on the command line. Corrects the help text to
"years" instead of "days".