regress: Reduce the default number of cores used
Reduces the default number of MCMC and OpenBlas threads to 1 for both.
De-obfuscates the number of threads and should make it easier to use on
a cluster with poor-man's parallelization as is, just using a single cpu
by default.
Anything else can still be set explicitly on the command line.