regress: Enable train/test split of the data
In Machine Learning techniques, the data is usually split into a
training set to fit the model parameters, and a test set against the
model is tested and evaluated. Introduces a command line option for the
train vs test fraction. We don't randomize (yet) and use the full data
set for plotting.