regress: Recompute the GP model before plotting
We plot the model samples together with all data and we need to
recompute the decomposition of the covariance matrix based on the full
data set once before. We also move the samples plotting out of the
args.mcmc guard so that we can run all plot routines in the same way.