regress: Remove data from mean model
We kept the data within the mean model for easier access and for
plotting, but they are not used to compute anything. Here we remove the
data for a better separation and to enable the model to be initialized
without any data input, just by setting the parameters.

Adapts the plotting helper functions and the calls to take the data as
arguments instead of relying on them being available via mean model
attributes.

Relies on the internally set times for the mcmc diagnostic predictions.