regress: Fix MCMC model times for residuals
George and celerite have different internal names for the time (resp. x)
values of the Gaussian Process (GP) kernels. We need them to calculate
the (GP) model values at the actual data points to evaluate the goodness
of the fit. Sets the model (data) times based on the GP model's
attribute (_x for `george` and _t for `celerite`).