regress: Pass data uncertainties to MCMC
Passing the data uncertainties to the MCMC subroutine enables us to
evaluate the goodness of the fit compared to the trivial model (mean and
standard deviation of the data).

Adjusts the call in `main()` accordingly and adds the keywords to the
keyword arguments to avoid unintentional side-effects when changing the
argument signature of the function.