regress: Clean up __main__
The shebang is not needed and was left over from when we used the file
as a standalone script. Delays importing astropy.time.Time and corner
until they are actually needed, gets rid of these dependences just for
running the script. Switches to normal numpy instead of autograd since
we don't (yet) use autograd's features.