regress: Fix R_Sun-Earth calculation
astropy.time.Time's `to_datetime()` fails on leap seconds. We use it
because of the different format options but we only need the day of the
year for the distance calculation. Replaces `to_datetime()` by setting
the format and extracting the day from that.

Changes the default time format to "jyear" to be consistent with the
data loading interface.