regress: Extract MCMC sample and test statistics
Puts the statistic logging from `mcmc_sample_model` into its own
function to make the code more readable and less complex. It also has
the advantage that we can feed it different test data and see how the
model performs with "out-of-bag" data, i.e. on data that were not
included in training (fitting) the model (parameters).