regress: Fix excluding SPEs and season selection
For SPEs, we look for all times in the data that are within the SPE
period and drop those, removing the need to reindex the data.
For seasonal selection, we `.load()` the dataset after `xr.concat()`,
otherwise `.where(..., drop=True)` fails with multiple conditions put
together with bitwise_and. Strangely, this happens only when using
`xr.concat()`.