readme: Update install procedure
Be more verbose about the install procedure. Although all dependencies
should be installed when running `pip`, it may be a good idea to install
some of them before separately, in particular when using `conda`.
We therefore list the main dependencies with a short note about what
they are used for.

[skip ci]