setup: Require some minimum versions
Requires minimal versions of numpy, scipy, and matplotlib to avoid
install and test failures.

The sciapy.plot_setup uses matplotlib's constrained layout which was
introduced in matplotlib version 2.2. The proxy interpolation with
scipy.interpolate.interp1d sets the `fill_value` to "extrapolate" which
was introduced in scipy version 0.17.0, and xarray raises a
numpy.AxisError which was introduced in numpy version 1.13.0.