docs: Update regression model intro notebook
Simplifies the model intro by removing all the bounds from the proxy
parameters and setting the parameters by hand first. This part is
to demonstrate how to use it.

Includes a second evaluation where we take the parameters from the fits,
but we access the file online instead of relying on having a local copy.

[skip ci]