docs: Install ipython on readthedocs
Hopefully fixes the display of the converted notebook(s). The provided
'rst' standard template for `nbconvert` puts the input cells as
`ipython3` and pygments (the syntax highlighter) does not know how to
handle these inputs unless `ipython` is installed.

Adds `ipython` to the readthedocs dependencies which hopefully brings
the input cells back.

[skip ci]