regress: Speed up proxy interpolation
We used the extrapolation of proxy values beyond the given ones.
However, `scipy` uses its own implementation in that case which is
slower than the `numpy` interpolation. So we switch off extrapolation
but still allow out-of-bounds values (giving NaNs) which speeds up the
interpolation considerably. We also swap the dimensions in the lifetime
correction such that the sum is over the last dimension (hopefully
C-order) and gives a little extra speed.