regress: Swap fit and burn-in order
This is a more cosmetic but nevertheless logic change. If the starting
point has not been optimized yet, we ran a so-called second burn-in.
However, it is more logical then to call the first burn-in a fit run,
and the second one burn-in.

In addition to that, we set the first random state explicitly to
`numpy.random.get_state()` which enables us to get reproducible results
by fixing the random seed.