regress: Add DIC to MCMC info output
Adds a few DIC (deivance information criterion)
https://doi.org/10.1111/j.1467-9574.2005.00278.x
outputs to the end of the MCMC run. Still not sure, which one to use,
apparently one should use the posterior mean, but the mode or median
should work similarly. The mode always gives pd = 0.

Fixes the long lines in the evidence log output.