regress mcmc: Use list for blobs
Wraps the log_likelihood we return as a blob with the probability
functions as a list element. In the future we may expand that list to
return more fancy stuff to the sampler. When calculating the IC from
the likelihoods, we need to pick out the correct column and correct
the type to `float`, otherwise `numpy.nanmean()` tends to fail.