regress: MSLL and more statistics
Adds more statistics to the model evaluation at the end. We use the
trained model to predict the test data, which means we don't `compute()`
again and cannot use the regular `likelihood()` function but have to
make up our own predictive probability density.
From that we calculate the "mean standardized log loss" as described in
R&W, 2006, section 2.5 (2.34).

Also adds "generalized R^2" scores by comparing the model prediction
with a prediction using the training data's mean and variance. Includes
several options for the test variance for now, using the provided test
errors (squared), using the variance from `predict()`, and using the
predictive variance plus the provided errors (noisy targets).

To avoid calculating the statistics twice, we only report the statistics
on the test data set, if the train or test fraction is actually smaller
than one.