regress: Make train_data optional for WAIC and LOO
The training data is only used if the `GP.predict()` method is used to
evaluate the MCMC samples. We use the likelihood by default and the
`train_data` are not needed in that case.

Sets the likelihood method unconditionally and overrides it only if
requested and `train_data` are provided. Should fix scrutinizer
complaining about the possibly undefined variable.