regress: Pass kwargs instead of 'args' to model
To simplify the (proxy) model setup, we put that into a separate
function and passed the `args` from `argparse`. This is (was) not very
flexible when used outside `main()` where `args` may not be available.
In the long run, we want to provide a convenience function with all the
parameters set to the defaults. Here we convert the `argparse` `args` to
`kwargs` and extract the setup from that which makes that function
already much more useful.

We put the proxy value normalization and everything connected to the
proxy name back to `main()` to keep the setup function simpler.