regress: Re-factor model setup
Puts the setup of the (trace gas) model from `__main__` to the model
submodule together with a default setup. Introduces convenience
functions to setup the model and makes them accessible from the outside.
Instead of setting up all in the main interface, we delegate the setup
to the new setup, changing the function call to a dictionary that
handles the various proxy probabilities.

Renames the combined model class to have a more descriptive name which
reflects the intention better. Updated the tests to reflect the name
change.