regress: Fix un-optimized MCMC position setup
In the non-optimized initial case for the MCMC sampler, we use the latin
hypercube sampling from `scipy.optimize`'s
`DifferentialEvolutionSolver`. This method initializes (our walkers)
only as a multiple of the number of parameters. In general, the number
of walkers may not be a multiple of the parameter vector's dimension.
We fill up the rest with samples from a normal distribution around the
initial parameter vector.