regress: Simplify lifetime adjustment for proxies
Delegates the lifetime correction for proxies to its own method and
introduces one for the gradient. Goes back to the loop solution instead
of broadcasting when summing up the last N proxy values. This may be a
bit slower for small data sets, but is faster for large ones since it
saves the creation of possibly large intermediate 2d arrays just to be
used for `numpy.sum()`.

Uses `numpy.interp()` instead of scipy, previously we assumed that using
`scipy`'s `interp1d` class would be cleaner and quicker than keeping the
times and values within our own class. Since `scipy`'s class does the
same but internally, we might as well use `numpy.interp()` ourselves.

Adds the possibility to use proxies with a different time resolution,
but we fixed it to one day for now.