regress: Fix combined proxy model for scalar input
All the submodels take care that the time input is `atleast_1d`,
but we initialized the empty result vector with the shape of the
original input. This shape leads to problems when summing up the
individual model results that have now a different shape.

Fixes this inconsistency by converting the time input to 1d before
initializing the result vector.