regress: Drop pandas for proxy reading
Uses `numpy.genfromtxt()` to read the proxy tables. The `pandas` people
are changing the API, and using it for the simple text tables was
probably overkill in the first place. To keep our code hopefully more
future-proof, we use plain `numpy` which gets the job done equally well.

`load_solar_gm_table()` now returns a structured array, so we don't need
`.values` anymore to get the proxy ndarray.