regress: Set encoding for newer numpy genfromtxt()
Fixes a future warning and hopefully helps us fixing Python 3.4
compatibility (on travis). It seems that older `astropy` versions choke
on the bytes that newer `numpy` version produce when the encoding is not
set (the default is 'bytes'). Setting the encoding to None should give
us more sensible (unicode) strings.