regress [ci]: Fix missing dev packages (Py3.4)
Travis' python 3.4 setup is missing the hdf5 and netcdf headers needed
by `netCDF4`. We install them before we set up the build environment.

Speeds up the builds a bit by fixing `numpy` to the minimum required
version. It does not need to be re-installed if an appropriate version
is already available.