regress: Fix zonal mean read with empty dataset
In cases where the selected latitudes or times returned an empty
dataset, `astropy` versions < 3.2.2 failed to convert the empty list to
`Time` and then to Julian epochs, raising an `IndexError`.
We fix this by testing for the returned dataset being empty and in that
case return an empty array. Works also with newer `astropy` versions which
can deal with empty arrays.

Adapts the *summerNH* test to look for empty arrays instead of a
raised exception.