docs: Fix level1c docs for RTD
Links the level1c README in the `docs/` subdir to make the link from the
main README work within the docs when not on github. Changes the case
for the main README to fix linking back from the level1c README.

Includes `sphinx_markdown_tables` to render markdown tables correctly.

Uses absolute (github) links for the scripts, relative paths work
correctly only locally and on github.

[skip ci]