lvl2: Netcdf3 Py2/3 compatibility
Scipy.io.netcdf / pupynere seems produces `bytes` instead of `str`
when reading netcdf3 with Python 3. This messes up the round trips of
the string to and from netcdf. Tries to decode the read strings to
ensure that the object has strings and matches the inputs.

This change only affects the internal handling and should normally not
be visible outside.