l2 pp: Fix bug in longitude, increase dev version
The longitude conversion from 0...360 to -180...180 was wrong, resulting
in wrong AACGM latitudes and longitudes. Fixes the conversion and
updates the test files with the correct geomagnetic coordinates.

Increases the development version in the main file and the level-2 post
processing test files to keep track of the changed geomagnetic
coordinates.