lvl2 igrf: Fix GM coordinate calculation
Fixes the angle (longitude) calculation of the centered-dipole and
eccentric dipole pole locations.

Laundal and Richmond, Space Sci. Rev. (2017) 206:27--59,
doi:10.1007/s11214-016-0275-y, note on p. 31:
"The Earth’s field is such that the dipole axis points roughly
southward, so that the dipole North Pole is really in the Southern
Hemisphere (SH).
However convention dictates that the axis of the geomagnetic dipole is
positive northward, hence the negative sign in the definition of mˆ."
Hence Phi_N in their Eq. (14) is actually Phi_S.

Fixes the longitude angle of the geodetic dipole offset in spherical
coordinates, `atan2()` without any shifts (Fraser-Smith 1987, Eq. (24)).

Notes the equations for reference.