lvl2 pp: Disable `astropy`'s time data download
Older python and astropy versions use a non-functional url to update the
leap second data, described on
https://docs.astropy.org/en/stable/utils/iers.html

Here the packaged data should be good enough, so we disable updating
those data in the post-processing scripts. This also prevents
downloading the data each time when testing the scripts.
But we keep a mirror url just in case so that the data can be activated
when really needed.

Should fix Python 2.7 and 3.5 ci.