regress: Initial `theano`/`pymc3` model interface
Implements model classes for sampling with `pymc3` including harmonic
terms and a proxy model with a variable lifetime.

`Pymc3` has a more sophisticated and cleaner interface when it comes to
probabilistic modelling and posterior sampling, but the `celerite` model
interface is incompatible with it.

The models have a rather limited set of parameters, the only caveat is
the sampling of previous values of the proxies. The convenience setup
functions are a relict from the `celerite` interface and may go away in
future versions because using the classes in a `pymc3.Model()` context
is more feasible than guessing default values.