doc: Update links in nbsphinx prologue
Links to the main notebook on github instead of the copy on rtd.
For some reason this is easier than trying to fix the duplicate
directory name when using "local" links.

Also updates the binder and nbviewer links, using docname seems to
include the subdirectory already. Those links still refer to the master
branch, so they will not work until the nbsphinx branch is merged and
the tutorial notebooks are actually in those places.

[skip ci]