docs: Use `jobs` to skip upgrade install of deps
Moves rtd's install step to a `job` with custom options.

Pins the numpy version so that importing the `theano` interface
for rendering the documentation succeeds. Installs the local package
without any upgrade option to not re-install newer numpy or other
packages with pip that are not really needed to build the documentation.
This should reduce the traffic for the doc building process because less
downloads are required.

[skip ci]