tests: Test "proxy" model
Implements a simple test for the "proxy" model interface.
Currently it generates some synthetic data and checks that we get back
what we put in, approximately, because of the MCMC sampling.

Intended to make sure that we get reproducible and similar results also
in future versions when changing the internal structure.