regress: Use PyMC3 "submodel"s for proxy setup
Uses a PyMC3 "submodel" for the individual proxies. This automatically
prepends the "name" to the individual random variables. Thus, no need to
"format" ourselves. Cleans up the code a bit.

The supposedly internal "log" variables have different names now,
but this interface is not finalized in any way just yet.