regress: Short-cut lookup without interpolation
Interpolation might not always be necessary or required/desired.
Implements short-cutting the (linear) interpolation step by
forward-filling the proxy values. Passed through via a keyword argument,
defaulting to doing the interpolation to not change the default
behaviour.
Interpolates nonetheless for the convolution step with a finite
(exponential) lifetime. To be more correct, we would have to
rewrite that convolution step to account for unequal spacing in the
proxy sampling.