tests: Pass test "values" instead of generating
Restructures the test-data generation by passing the data through the
proxy time delay and distortion with a finite lifetime, instead of
generating a second one within that function.

We also set the frequency, fixed to 1 (year) for now; and we set the
proxy's delta-t to be more flexible in the future.