tests: Use mjd for mock-time in pymc mcmc tests
Uses mock time series in (modified Julian) days for the pyMC mcmc tests.
We convert them to fractional years, shifting them so that using a
frequency of 1 y or 365.25 both start at "zero".
We do that internally within the proxy class, but here we have to shift
manually so that both versions (would) produce the same mock data array.

We loosen the comparison constrains a little because they are a little
further apart now, but we still want the tests to pass.