regress pymc: Fix fracyear vs (m)jd time decision
Fixes the way we determine if we are given (Julian) day numbers or
fractional years as the proxy's time argument. Since the class members
are tensors, a simple comparison with floats or integers does not give
the desired results. Uses `theano.tensor`'s `switch` to compare them and
set the result accordingly.