regress cli: Rewrite OpenMP/OpenBLAS thread setup
Sets/changes the environment variables for OpenMP/OpenBLAS thread
settings instead of using numpy and calling the library function.
First, relying on OpenBLAS to be linked is too much here and may not
have the desired effect. Second, `numpy` is deprecating the `distutils`
interface, and it is not really needed. If folks want to change the
number of threads used for the mcmc fits, they can also do that by
setting the necessary environment variables themselves.